Simpilfy comment event
